+++ tests/Feature/Console/DebugModule/Fixtures/ImperativeModule/ImperativeModuleFacade.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Console\DebugModule\Fixtures\ImperativeModule;
+
+use Gacela\Framework\AbstractFacade;
+
+final class ImperativeModuleFacade extends AbstractFacade
+{
+}